GIC IRQ Done. (Tested on CM4)

+++ kernel/hardware/irq/bcm2837_irq_manager.hpp
@@ -5,9 +5,9 @@
 namespace BCM2837_IRQManager {
 void init();
 
-void enable_irq(uint64_t irq_id);
-void disable_irq(uint64_t irq_id);
+void enable_irq(uint32_t irq_id);
+void disable_irq(uint32_t irq_id);
 
-void mask_as_processed(uint64_t irq_id);
-bool has_pending_interrupt(uint64_t* irq_id);
+void mask_as_processed(uint32_t irq_id);
+bool has_pending_interrupt(uint32_t* irq_id);
 };  // namespace BCM2837_IRQManager
